Feat: Added API tests jest&supertest
